--- a/package.json
+++ b/package.json
@@ -35,6 +35,7 @@
   },
   "dependencies": {
     "fast-decode-uri-component": "^1.0.0",
+    "generate-radix-tree": "^1.0.2",
     "safe-regex": "^1.1.0",
     "semver-store": "^0.3.0"
   }