@mcollina 2018-09-06
>  Fix regex issue. I'm not sure how your regex matcher works. I just do an eager match against the pattern passed, but you support $ as an inplace regex which I'm unsure what means

These are express styles regexps, with parameters. such as:

```
Route path: /flights/:from-:to
Request URL: http://localhost:3000/flights/LAX-SFO
req.params: { "from": "LAX", "to": "SFO" }
```

But also

```
Route path: /user/:userId(\d+)
Request URL: http://localhost:3000/user/42
req.params: {"userId": "42"}
```

Turn out that it is one of the most depended-on express features.

---

cc @allevo you played with this, you might want to help out1