@allevo 2018-09-06
> cc @allevo you played with this, you might want to help out1

I'd like to be focused on [the router I'm writting](https://www.npmjs.com/package/road-marking) that seems to be a lot faster that `find-my-way`.

Anyway all supported feature should be kept even if the performance decreases.