@mafintosh 2018-09-06
@mcollina does the regex only match until the next '/'? That was the part I was unsure about