@mcollina 2018-09-06
yes, it does.
Il giorno gio 6 set 2018 alle 13:15 Mathias Buus <notifications@github.com>
ha scritto:

> @mcollina <https://github.com/mcollina> does the regex only match until
> the next '/'? That was the part I was unsure about
>
> —
> You are receiving this because you were mentioned.
>
>
> Reply to this email directly, view it on GitHub
> <https://github.com/delvedor/find-my-way/pull/89#issuecomment-419057090>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AADL46IeDneW-wOUFR71a5j0taVCj2Lsks5uYQPBgaJpZM4Wb-ue>
> .
>